requirements.txt
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.30.0
plotly>=5.18.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
scikit-learn
openpyxl